+++ app/Controllers/RoomType.php
@@ -2,6 +2,7 @@
 
 namespace App\Controllers;
 use App\Models\RoomTypeModel;
+use App\Controllers\PermissionChecker;
 
 class RoomType extends BaseController
 {
@@ -9,11 +10,12 @@ class RoomType extends BaseController
         helper('form');
 		$this->roomTypeModel = new RoomTypeModel();
         $this->session = \Config\Services::session();
+        $this->permcheck = new PermissionChecker();
     }
 
 	public function index()
 	{
-        if($this->session->has('logged_in')){
+        if($this->session->has('logged_in') && $this->permcheck->check($this->session->get('id'),'RoomType')){
             date_default_timezone_set('Asia/Manila');
             $data['room_types'] = $this->roomTypeModel->findAll();
             return view('room/type/index', $data);
@@ -22,7 +24,7 @@ class RoomType extends BaseController
 	}
 
     public function add(){
-        if($this->session->has('logged_in')){
+        if($this->session->has('logged_in') && $this->permcheck->check($this->session->get('id'),'RoomTypeAdd')){
             date_default_timezone_set('Asia/Manila');
             if($this->request->getMethod()=='post'){
                 $data =[
@@ -45,7 +47,7 @@ class RoomType extends BaseController
     }
 
     public function edit($id=null){
-        if($this->session->has('logged_in')){
+        if($this->session->has('logged_in') && $this->permcheck->check($this->session->get('id'),'RoomTypeEdit')){
             date_default_timezone_set('Asia/Manila');
 
             $data['room_type'] = $this->roomTypeModel->find($id);
@@ -70,7 +72,7 @@ class RoomType extends BaseController
     }
 
     public function delete($id=null){
-        if($this->session->has('logged_in')){
+        if($this->session->has('logged_in') && $this->permcheck->check($this->session->get('id'),'RoomTypeDelete')){
             if($this->roomTypeModel->where('room_type_id',$id)->delete()){
                 $this->session->setTempdata('success', 'Deleted Successfully!', 3);
                 return redirect()->to(base_url().'/room/type');
